Component/Structure Improvements
*Changed structure of major view/control components.

Moved IP/port to be a permanent property of the store.

Moved ip manipulation functions to be in a single element (AutoGenDeviceAddress.tsx --><DeviceAddressManager>)

Moved all functions that input device address to now input a device address object which contains a sub object of the ip (4 fragments), and a port.

*Moved all logic out of the major views. Major views now only display the actual objects desired when set in render. Can pass props.

*TODO:

Need to add doctor and patietn modularity fixes to functional components/models.

___________________________________

Autogen is done.

AutogenView is now
*AutogenView
   (Contains)
*AutoGenControls
*AutoGenDemos
*AutoGenDeviceAddress