SignalR fixes + Endpoint. Added comparison and components json.
SignalR:

Added ActionStore action for requesting the connect-rtc for server (gets true/false depending on connection status of server to the signalr hub).

Added properties to the project to reflect server status.

Added call in component mount to initiate a connection for server to hub.

Autogen: Added a components (reference) json. This json is to be used by the gateway and served for comparison purposes. The json lists required and optional data properties as well as names, data types, and variants for components. (Component description json).

Added functions (in sandbox) and tested comparison of served data with the components json. Comparison finds matches of data properties, and compares the optional and required fields and unaccounted for fields, returning the best component match(es) for each data structure.

This can further be used to reference matches for requried fields only for editing the control used for  a model's view.