Minor Convention update (AutogenDemos -> S3ProjectsContainer)
Added dynamic, bootstrap compatible (start) for typing, and use.
AutogenDemos is now S3ProjectsContainer since it is a container for S3Projects :)
Added additional props and handling for defaults such as className, and defaultBucket.  This will allow us to reuse this component (potentially), or update for specific builds, or pages on the component declaration itself as opposed to within the component.
Available projects can now be hidden from the component level.